pyproject.toml
[tool.poetry]
name = "maria-mlops"
version = "0.1.0"
description = ""
authors = ["maria"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.5.1"
pandas = "^2.2.3"
torchvision = "^0.20.1"
torchsummary = "^1.5.1"
matplotlib = "^3.10.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
lightning = "^2.5.0.post0"
pytorch-lightning = "^2.5.0.post0"
mlflow = "^2.19.0"
dvc = "^3.59.0"
tensorboard = "^2.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
start = "poetry run python commands.py"
train = "poetry run python breeds_breeds/train.py"
test = "poetry run python breeds_breeds/test.py"
infer = "poetry run python breeds_breeds/infer.py"
